requirements.txt
aiogram==3.0.0b7
python-dotenv==1.0.0
requests==2.28.2
redis==4.5.4